+++ src/maths.h
@@ -2,5 +2,5 @@
 long long int suma (long long int numeroUno, long long int numeroDos);
 long long int resta (long long int numeroUno, long long int numeroDos);
 long long int multiplicacion (long long int numeroUno, long long int numeroDos);
-double division (long long int numeroUno, long long int numeroDos);
+double division (double numeroUno, double numeroDos);
 unsigned long long factorial(long long int numero);
